+++ federation/tests/utils/test_text.py
@@ -1,4 +1,4 @@
-from federation.utils.text import decode_if_bytes, encode_if_text, validate_handle, process_text_links
+from federation.utils.text import decode_if_bytes, encode_if_text, validate_handle, process_text_links, find_tags
 
 
 def test_decode_if_bytes():
@@ -9,6 +9,70 @@ def test_decode_if_bytes():
 def test_encode_if_text():
     assert encode_if_text(b"foobar") == b"foobar"
     assert encode_if_text("foobar") == b"foobar"
+
+
+class TestFindTags:
+    @staticmethod
+    def _replacer(text):
+        return f"#{text}/{text.lower()}"
+
+    def test_all_tags_are_parsed_from_text(self):
+        source = "#starting and #MixED with some #line\nendings also tags can\n#start on new line"
+        tags, text = find_tags(source)
+        assert tags == {"starting", "mixed", "line", "start"}
+        assert text == source
+        tags, text = find_tags(source, replacer=self._replacer)
+        assert text == "#starting/starting and #MixED/mixed with some #line/line\nendings also tags can\n" \
+                       "#start/start on new line"
+
+    def test_code_block_tags_ignored(self):
+        source = "foo\n```\n#code\n```\n#notcode\n\n    #alsocode\n"
+        tags, text = find_tags(source)
+        assert tags == {"notcode"}
+        assert text == source
+        tags, text = find_tags(source, replacer=self._replacer)
+        assert text == "foo\n```\n#code\n```\n#notcode/notcode\n\n    #alsocode\n"
+
+    def test_endings_are_filtered_out(self):
+        source = "#parenthesis) #exp! #list]"
+        tags, text = find_tags(source)
+        assert tags == {"parenthesis", "exp", "list"}
+        assert text == source
+        tags, text = find_tags(source, replacer=self._replacer)
+        assert text == "#parenthesis/parenthesis) #exp/exp! #list/list]"
+
+    def test_finds_tags(self):
+        source = "#post **Foobar** #tag #OtherTag #third\n#fourth"
+        tags, text = find_tags(source)
+        assert tags == {"third", "fourth", "post", "othertag", "tag"}
+        assert text == source
+        tags, text = find_tags(source, replacer=self._replacer)
+        assert text == "#post/post **Foobar** #tag/tag #OtherTag/othertag #third/third\n#fourth/fourth"
+
+    def test_postfixed_tags(self):
+        source = "#foo) #bar] #hoo, #hee."
+        tags, text = find_tags(source)
+        assert tags == {"foo", "bar", "hoo", "hee"}
+        assert text == source
+        tags, text = find_tags(source, replacer=self._replacer)
+        assert text == "#foo/foo) #bar/bar] #hoo/hoo, #hee/hee."
+
+    def test_prefixed_tags(self):
+        source = "(#foo [#bar"
+        tags, text = find_tags(source)
+        assert tags == {"foo", "bar"}
+        assert text == source
+        tags, text = find_tags(source, replacer=self._replacer)
+        assert text == "(#foo/foo [#bar/bar"
+
+    def test_invalid_text_returns_no_tags(self):
+        source = "#a!a #a#a #a$a #a%a #a^a #a&a #a*a #a+a #a.a #a,a #a@a #a£a #a/a #a(a #a)a #a=a " \
+                 "#a?a #a`a #a'a #a\\a #a{a #a[a #a]a #a}a #a~a #a;a #a:a #a\"a #a’a #a”a #\xa0cd"
+        tags, text = find_tags(source)
+        assert tags == set()
+        assert text == source
+        tags, text = find_tags(source, replacer=self._replacer)
+        assert text == source
 
 
 class TestProcessTextLinks: